Fixed MissingMethodException, also fixed the mixed up open/close actions
