+++ Source/SubWall_JobDefOf.cs
@@ -1,0 +1,14 @@
+﻿using RimWorld;
+using Verse;
+
+namespace SubWall
+{
+    [DefOf]
+    internal class SubWall_JobDefOf
+    {
+        public static JobDef RetDef_CloseGate;
+        public static JobDef RetDef_OpenGate;
+        public static JobDef RetDef_ClosePort;
+        public static JobDef RetDef_OpenPort;
+    }
+}
